Conflicts:
	Projects/AssetsManagement/Src/Globals.h
	Projects/ClientApp/Src/ClientApp.vcxproj
	Projects/ClientApp/Src/ClientApplication.h
	Projects/Gameplay/Src/Gameplay.vcxproj
	Projects/Gameplay/Src/Gameplay.vcxproj.filters
	Projects/Graphics/Src/BufferFactory.cpp
	Projects/Graphics/Src/BufferFactory.h
	Projects/Graphics/Src/Graphics.vcxproj.filters
	Projects/Graphics/Src/GraphicsWrapper.cpp

+++ Projects/Gameplay/Src/InputBackendSystem.cpp
@@ -208,7 +208,7 @@ vector<int>* InputBackendSystem::vectorFromDeviceType( InputHelper::INPUT_DEVICE
 
 void InputBackendSystem::expandIdxVectorIfNecessary( vector<int>* p_vec, int p_idx )
 {
-	if( p_vec->size() <= p_idx )
+	if( p_vec->size() <= ( unsigned int ) p_idx )
 	{
 		p_vec->resize(p_idx+1);
 	}
